+++ SD85_WebBookOnline.Share/Models/CartItems.cs
@@ -15,9 +15,9 @@ namespace SD85_WebBookOnline.Share.Models
         public Guid? ComboID { get; set; }
         public Guid? BookID { get; set; }
         public string ItemName { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Quantity   { get; set; }
-        public double ToTal { get; set; }
+        public decimal ToTal { get; set; }
         public int Status { get; set; }
 
         public virtual Combo? Combo { get; set; }
